+++ docs/source/changelog.rst
@@ -7,6 +7,7 @@ Version 0.156.0
 ===============
 
 - Speed up environment loading for big scenes with lots of static links.
+- Deprecate restoring of grabbed bodies by `KinBodyStateSaver` from one env to another env since it's hard to restore any kind of bodies into different env. Instead, use the dedicated private API in `Environment::_Clone`.
 
 Version 0.155.0
 ===============